Refactor to System.Diagnostics.Metrics.Temp package
