+++ src/System.Diagnostics.Metrics.Temp/Instrument.cs
@@ -18,7 +18,6 @@ using System.Collections.Concurrent;
 using System.Collections.Generic;
 
 #nullable enable
-#pragma warning disable SA1623, SA1611, SA1615
 
 namespace System.Diagnostics.Metrics
 {
