+++ src/System.Diagnostics.Metrics.Temp/ObservableInstrument.cs
@@ -17,7 +17,6 @@
 using System.Collections.Generic;
 
 #nullable enable
-#pragma warning disable SA1623, SA1611, SA1615
 
 namespace System.Diagnostics.Metrics
 {
