+++ src/System.Diagnostics.Metrics.Temp/System.Diagnostics.Metrics.Temp.csproj
@@ -3,7 +3,7 @@
     <TargetFrameworks>netstandard2.0;net452</TargetFrameworks>
     <Description>Temporary .NET Metrics API</Description>
     <RootNamespace>OpenTelemetry</RootNamespace>
-    <NoWarn>$(NoWarn),CS0618, CS1591</NoWarn>
+    <NoWarn>,1573,1712,CS0618, CS1591;NU5105;105;NU5105;NU5105, SA1611; SA1615; NU5105;NU5105; SA1625;NU5105; SA1625;;NU5105;SA1623;</NoWarn>
     <TreatWarningsAsErrors>false</TreatWarningsAsErrors>
     <IsPackable>false</IsPackable>
   </PropertyGroup>
